init_prog: launch submit module with num-jobs argument
avoids the hangs we were experiencing when the number of jobs in the
input file exceeds the number of jobs passed in at the command
line. For example, if the user requests a simulation of 10K jobs but
passes in a file with 500K jobs, the initial_program will terminate
after 10K jobs complete but the broker will hang indefinitely.